% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_utils.R
\name{cPOST}
\alias{cPOST}
\title{cPOST}
\usage{
cPOST(
  args,
  fun_url,
  open_cpu_url = options("open_cpu_url"),
  return_value = TRUE
)
}
\arguments{
\item{args}{named string which will be converted to 'name1=value1&name2=value2'}
}
\description{
cPOST
}
\details{
pass POST to system curl command. Useful for passing unquoted (unquotable?) opencpu session object references.
}
