% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul_utils.R
\name{get_ocpu_list_item}
\alias{get_ocpu_list_item}
\title{get_ocpu_list_item}
\usage{
get_ocpu_list_item(connection, body)
}
\arguments{
\item{body}{}
}
\value{
named element from objects created in other sessions
}
\description{
get_ocpu_list_item
}
