% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul_utils.R
\name{get_ocpu_obj}
\alias{get_ocpu_obj}
\title{get_ocpu_obj}
\usage{
get_ocpu_obj(session, open_cpu_url = options("open_cpu_url"))
}
\description{
get_ocpu_obj
}
\details{
get ocpu results based on session
}
