% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocpu_call}
\alias{ocpu_call}
\title{ocpu_call}
\usage{
ocpu_call(
  fun,
  body = NULL,
  pkg_url = NULL,
  base_url = getOption("open_cpu_url"),
  ...
)
}
\description{
ocpu_call
}
\details{
call function and serialize results to json
}
