% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul_utils.R
\name{session_list_item}
\alias{session_list_item}
\title{session_list_item}
\usage{
session_list_item(obj, name)
}
\description{
session_list_item
}
\details{
get list element by name from obj created in a previous session
}
