% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocpu_post}
\alias{ocpu_post}
\title{ocpu_post}
\usage{
ocpu_post(
  fun,
  body,
  pkg_url = getOption("dave.stat_url"),
  base_url = getOption("open_cpu_url")
)
}
\description{
ocpu_post
}
\details{
can get fails to memoize if called from a shiny app
}
